REF: Refactor DBN versioning
